+++ src/Aws/StorageGateway/Resources/storagegateway-2013-06-30.php
@@ -660,8 +660,6 @@ return array (
                     'required' => true,
                     'type' => 'numeric',
                     'location' => 'json',
-                    'minimum' => 107374182400,
-                    'maximum' => 2748779069440,
                 ),
                 'ClientToken' => array(
                     'required' => true,
@@ -1847,6 +1845,46 @@ return array (
                 ),
             ),
         ),
+        'ResetCache' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'ResetCacheOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'StorageGateway_20130630.ResetCache',
+                ),
+                'GatewayARN' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 50,
+                    'maxLength' => 500,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An exception occurred because an invalid gateway request was issued to the service. See the error and message fields for more information.',
+                    'class' => 'InvalidGatewayRequestException',
+                ),
+                array(
+                    'reason' => 'An internal server error has occurred during the request. See the error and message fields for more information.',
+                    'class' => 'InternalServerErrorException',
+                ),
+            ),
+        ),
         'RetrieveTapeArchive' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -2327,6 +2365,53 @@ return array (
                     'location' => 'json',
                     'minLength' => 1,
                     'maxLength' => 255,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An exception occurred because an invalid gateway request was issued to the service. See the error and message fields for more information.',
+                    'class' => 'InvalidGatewayRequestException',
+                ),
+                array(
+                    'reason' => 'An internal server error has occurred during the request. See the error and message fields for more information.',
+                    'class' => 'InternalServerErrorException',
+                ),
+            ),
+        ),
+        'UpdateVTLDeviceType' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'UpdateVTLDeviceTypeOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'StorageGateway_20130630.UpdateVTLDeviceType',
+                ),
+                'VTLDeviceARN' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 50,
+                    'maxLength' => 500,
+                ),
+                'DeviceType' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 2,
+                    'maxLength' => 50,
                 ),
             ),
             'errorResponses' => array(
@@ -3139,6 +3224,9 @@ return array (
                             'DiskNode' => array(
                                 'type' => 'string',
                             ),
+                            'DiskStatus' => array(
+                                'type' => 'string',
+                            ),
                             'DiskSizeInBytes' => array(
                                 'type' => 'numeric',
                             ),
@@ -3212,6 +3300,15 @@ return array (
                             ),
                         ),
                     ),
+                ),
+            ),
+        ),
+        'ResetCacheOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'GatewayARN' => array(
+                    'type' => 'string',
                 ),
             ),
         ),
@@ -3316,6 +3413,15 @@ return array (
                 'VolumeARN' => array(
                     'type' => 'string',
                     'location' => 'json',
+                ),
+            ),
+        ),
+        'UpdateVTLDeviceTypeOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'VTLDeviceARN' => array(
+                    'type' => 'string',
                 ),
             ),
         ),